requirements.txt
absl-py==0.9.0
astor==0.7.1
astunparse==1.6.3
autopep8==1.5.3
cachetools==4.1.0
certifi==2020.6.20
chardet==3.0.4
gast==0.3.3
google-auth==1.18.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.23.0
h5py==2.10.0
idna==2.9
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.2
mmh3==2.5.1
mock==3.0.5
numpy==1.18.1
oauthlib==3.1.0
opt-einsum==3.2.1
protobuf==3.8.0
pyaml==19.4.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pygame==1.9.6
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
six==1.14.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
toml==0.10.1
urllib3==1.25.9
Werkzeug==1.0.0
wrapt==1.12.1
